Laravel Download Model
The Download Model allows the user to download a .CSV file with all of the information they have selected on the website. The Download class also has methods which check if that file already exists, a method which saves information on which user downloaded which data, it also provides information on the which business the user belongs to, and a few other things.