Laravel Web Routes
The Web Routes file contains all the routes the users are able to  navigate through. The paths are separated by the amount of access that is needed for the user to view it. Each group contains middleware that checks whether the user has right amount of privilege in order to view the file.