Laravel Subscriber Model
The Subscriber Model allows the user to subscribe to our service and immediately stores the user's information to our database. It also send the user an email with a URL that includes a unique hash in order for the user to confirm their email as well as prevent other people from using email which are not theirs.